Rename Python package to  and updated files to match.
