+++ MANIFEST.in
@@ -1,3 +1,4 @@
 include README.md
-include LICENSE.txt
+include LICENSE
+include NOTICE
 recursive-exclude examples *
