+++ aws_cli_migrate/rules/__init__.py
@@ -10,6 +10,6 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-from awsclilinter.rules.base import LintFinding, LintRule
+from aws_cli_migrate.rules.base import LintFinding, LintRule
 
 __all__ = ["LintRule", "LintFinding"]
