+++ pyproject.toml
@@ -3,13 +3,13 @@ requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "awsclilinter"
+name = "aws-cli-migrate"
 version = "1.0.0"
 description = "Command-line tool to lint and modify bash scripts that use AWS CLI v1 to prevent breakage from changes introduced in AWS CLI v2."
 readme = "README.md"
 requires-python = ">=3.9"
 license = "Apache-2.0"
-keywords = ["aws", "cli", "linter", "bash", "script", "migration", "v1", "v2"]
+keywords = ["aws", "cli", "linter", "bash", "script", "migration", "v1", "v2", "codemod", "migrate"]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Intended Audience :: Developers",
@@ -42,7 +42,7 @@ Documentation = "https://github.com/aws/aws-cli"
 "Source Code" = "https://github.com/aws/aws-cli"
 
 [project.scripts]
-migrate-aws-cli = "awsclilinter.cli:main"
+migrate-aws-cli = "aws_cli_migrate.cli:main"
 
 [tool.ruff]
 line-length = 100
@@ -63,7 +63,7 @@ python_functions = ["test_*"]
 
 [tool.setuptools.packages.find]
 where = ["."]
-include = ["awsclilinter*"]
+include = ["aws_cli_migrate*"]
 
 [[tool.uv.index]]
 name = "testpypi"
